# datascope-api/src/main/resources/swagger/query-history-api.yaml
openapi: 3.0.0
info:
  title: 查询历史API
  description: |
    # 查询历史管理API
    
    提供查询历史的查询、统计和管理功能。
    
    ## 主要功能
    - 查询历史记录的增删改查
    - 按用户查询历史记录
    - 查询执行统计信息
    - 慢查询分析
    - 失败查询分析
    
    ## 使用说明
    1. 所有时间参数使用 `yyyy-MM-dd HH:mm:ss` 格式
    2. 分页参数默认值：pageNum=1, pageSize=10
    3. 响应格式统一使用标准Response封装
    
    ## 错误码说明
    - 200: 成功
    - 400: 请求参数错误
    - 401: 未授权
    - 403: 禁止访问
    - 404: 资源不存在
    - 500: 服务器内部错误
  version: 1.0.0
  contact:
    name: DataScope Team
    email: support@datascope.com
    url: https://datascope.com

servers:
  - url: /api/v1
    description: API版本1

tags:
  - name: query-histories
    description: 查询历史管理

paths:
  /query-histories/{id}:
    get:
      tags:
        - query-histories
      summary: 获取查询历史详情
      description: |
        根据ID获取单条查询历史记录的详细信息。
        
        ## 使用场景
        - 查看特定查询的执行详情
        - 分析查询失败原因
        - 查看查询参数和结果
      parameters:
        - name: id
          in: path
          required: true
          description: 历史记录ID
          schema:
            type: string
          example: "1234567890"
      responses:
        '200':
          description: 成功返回历史记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryResponse'
              example:
                $ref: 'examples/query-history-examples.json#/getHistoryById'
        '404':
          description: 历史记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                code: "404"
                message: "历史记录不存在"
                data: null
    delete:
      tags:
        - query-histories
      summary: 删除历史记录
      description: |
        根据ID删除单条查询历史记录。
        
        ## 使用场景
        - 清理无用的历史记录
        - 删除敏感查询记录
      parameters:
        - name: id
          in: path
          required: true
          description: 历史记录ID
          schema:
            type: string
          example: "1234567890"
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                success: true
                code: "200"
                message: "删除成功"
                data: null
        '404':
          description: 历史记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                code: "404"
                message: "历史记录不存在"
                data: null

  /query-histories/by-query/{queryId}:
    get:
      tags:
        - query-histories
      summary: 获取查询相关的历史记录
      description: |
        获取指定查询ID的所有历史执行记录。
        
        ## 使用场景
        - 分析查询的执行情况
        - 查看查询的执行趋势
        - 优化查询性能
      parameters:
        - name: queryId
          in: path
          required: true
          description: 查询ID
          schema:
            type: string
          example: "query123"
      responses:
        '200':
          description: 成功返回历史记录列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryListResponse'

  /query-histories/user/{userId}:
    get:
      tags:
        - query-histories
      summary: 分页查询用户的查询历史
      description: |
        分页获取指定用户的查询历史记录。
        
        ## 使用场景
        - 查看用户的查询记录
        - 分析用户的查询习惯
        - 监控用户的查询行为
        
        ## 查询条件说明
        - 时间范围：可选，默认查询最近7天
        - 执行状态：可选，支持成功/失败过滤
        - 分页参数：必选，默认每页10条
      parameters:
        - name: userId
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
          example: "user123"
        - name: startTime
          in: query
          description: 开始时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-01 00:00:00"
        - name: endTime
          in: query
          description: 结束时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-15 23:59:59"
        - name: status
          in: query
          description: 状态
          schema:
            type: string
            enum: [SUCCESS, FAILED]
          example: "SUCCESS"
        - name: pageNum
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: pageSize
          in: query
          description: 每页大小
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
      responses:
        '200':
          description: 成功返回分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryPageResponse'
              example:
                $ref: 'examples/query-history-examples.json#/getUserHistoryByPage'

  /query-histories/stats:
    get:
      tags:
        - query-histories
      summary: 获取执行统计信息
      description: |
        获取指定时间范围内的查询执行统计信息。
        
        ## 统计指标
        - 总执行次数
        - 平均执行时间
        - 最大执行时间
        - 最小执行时间
        - 总影响行数
        - 成功次数
        - 失败次数
        
        ## 使用场景
        - 监控查询性能
        - 分析系统负载
        - 优化查询效率
      parameters:
        - name: startTime
          in: query
          required: true
          description: 开始时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-01 00:00:00"
        - name: endTime
          in: query
          required: true
          description: 结束时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-15 23:59:59"
        - name: userId
          in: query
          description: 用户ID
          schema:
            type: string
          example: "user123"
        - name: queryId
          in: query
          description: 查询ID
          schema:
            type: string
          example: "query123"
      responses:
        '200':
          description: 成功返回统计信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryStatsResponse'
              example:
                $ref: 'examples/query-history-examples.json#/getExecutionStats'

  /query-histories/failures:
    get:
      tags:
        - query-histories
      summary: 获取最近失败的查询记录
      description: |
        获取最近一段时间内执行失败的查询记录。
        
        ## 使用场景
        - 监控查询失败情况
        - 分析失败原因
        - 优化查询语句
      parameters:
        - name: userId
          in: query
          description: 用户ID
          schema:
            type: string
          example: "user123"
        - name: limit
          in: query
          description: 限制数量
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
      responses:
        '200':
          description: 成功返回失败记录列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryListResponse'

  /query-histories/slow-queries:
    get:
      tags:
        - query-histories
      summary: 获取慢查询记录
      description: |
        获取执行时间超过阈值的慢查询记录。
        
        ## 使用场景
        - 监控查询性能
        - 识别性能瓶颈
        - 优化慢查询
        
        ## 参数说明
        - thresholdMillis: 执行时间阈值，单位毫秒
        - startTime/endTime: 查询时间范围
        - limit: 返回记录数量限制
      parameters:
        - name: thresholdMillis
          in: query
          required: true
          description: 执行时间阈值（毫秒）
          schema:
            type: integer
            minimum: 1000
          example: 3000
        - name: startTime
          in: query
          required: true
          description: 开始时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-01 00:00:00"
        - name: endTime
          in: query
          required: true
          description: 结束时间 (yyyy-MM-dd HH:mm:ss)
          schema:
            type: string
            format: date-time
          example: "2024-03-15 23:59:59"
        - name: limit
          in: query
          description: 限制数量
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
      responses:
        '200':
          description: 成功返回慢查询记录列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryHistoryListResponse'
              example:
                $ref: 'examples/query-history-examples.json#/getSlowQueries'

components:
  schemas:
    QueryHistory:
      type: object
      description: 查询历史记录
      required:
        - id
        - queryId
        - executionTime
        - status
        - createdAt
        - createdBy
      properties:
        id:
          type: string
          description: 历史记录ID
          example: "1234567890"
        queryId:
          type: string
          description: 查询ID
          example: "query123"
        parameters:
          type: string
          description: |
            查询参数，JSON格式字符串
            包含查询条件、过滤条件等
          example: "{\"startDate\":\"2024-03-01\",\"endDate\":\"2024-03-15\"}"
        executionTime:
          type: integer
          description: 执行时间（毫秒）
          minimum: 0
          example: 1500
        affectedRows:
          type: integer
          description: 影响行数
          minimum: 0
          example: 100
        status:
          type: string
          description: 执行状态
          enum: [SUCCESS, FAILED]
          example: "SUCCESS"
        errorMessage:
          type: string
          description: 错误信息
          example: null
        executionIp:
          type: string
          description: 执行IP
          example: "192.168.1.100"
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-03-15 10:30:00"
        createdBy:
          type: string
          description: 创建者
          example: "admin"

    Response:
      type: object
      description: 统一响应格式
      required:
        - success
        - code
        - message
      properties:
        success:
          type: boolean
          description: 是否成功
          example: true
        code:
          type: string
          description: 响应码
          example: "200"
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          type: object
          description: 响应数据
          nullable: true

    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              type: object
              nullable: true

    QueryHistoryResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/QueryHistory'

    QueryHistoryListResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/QueryHistory'

    PageResponse:
      type: object
      description: 分页响应
      required:
        - records
        - total
        - pageNum
        - pageSize
      properties:
        records:
          type: array
          description: 记录列表
          items:
            type: object
        total:
          type: integer
          description: 总记录数
          minimum: 0
          example: 50
        pageNum:
          type: integer
          description: 当前页码
          minimum: 1
          example: 1
        pageSize:
          type: integer
          description: 每页大小
          minimum: 1
          maximum: 100
          example: 10

    QueryHistoryPageResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageResponse'
                - type: object
                  properties:
                    records:
                      type: array
                      items:
                        $ref: '#/components/schemas/QueryHistory'

    QueryHistoryStatsResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              type: object
              required:
                - total_executions
                - avg_execution_time
                - max_execution_time
                - min_execution_time
                - total_affected_rows
                - success_count
                - failed_count
              properties:
                total_executions:
                  type: integer
                  description: 总执行次数
                  minimum: 0
                  example: 1000
                avg_execution_time:
                  type: number
                  description: 平均执行时间
                  minimum: 0
                  example: 1200.5
                max_execution_time:
                  type: number
                  description: 最大执行时间
                  minimum: 0
                  example: 5000
                min_execution_time:
                  type: number
                  description: 最小执行时间
                  minimum: 0
                  example: 100
                total_affected_rows:
                  type: integer
                  description: 总影响行数
                  minimum: 0
                  example: 50000
                success_count:
                  type: integer
                  description: 成功次数
                  minimum: 0
                  example: 950
                failed_count:
                  type: integer
                  description: 失败次数
                  minimum: 0
                  example: 50